# Improve performance, especially on non-ASCII codepoints