… scripts/unicode.py
… src/tables.rs