@Manishearth 2022-06-10
I may not have time to review this soon but it looks promising!

Would it be possible to include the code for the benchmark, perhaps disabled in Cargo.toml, with instructions for downloading the file?