@mjguynn 2022-06-10
I've just added the benchmark code & instructions. Currently it will just benchmark against an empty string if it cant find the data files, let me know if you'd prefer configuration in Cargo.toml instead.